requirements.txt
pymongo
langchain-community
langchain-openai
tiktoken
tqdm
PyYAML
colorama
python-docx
flask